Extract API routines into helper functions
