+++ lib/ret_web.ex
@@ -29,8 +29,9 @@ defmodule RetWeb do
 
   def view do
     quote do
-      use Phoenix.View, root: "lib/ret_web/templates",
-                        namespace: RetWeb
+      use Phoenix.View,
+        root: "lib/ret_web/templates",
+        namespace: RetWeb
 
       # Import convenience functions from controllers
       import Phoenix.Controller, only: [get_flash: 2, view_module: 1]
